Receiving activity types is allocated to a separate controller
